+++ Tests/PuppyTests/FileLoggerTests.swift
@@ -49,13 +49,71 @@ final class FileLoggerTests: XCTestCase {
         let emptyFileURL = URL(fileURLWithPath: "").absoluteURL     // file:///private/tmp/
         XCTAssertThrowsError(try FileLogger("com.example.yourapp.filelogger.notfile0", fileURL: emptyFileURL)) { error in
             XCTAssertEqual(error as? FileError, .isNotFile(url: emptyFileURL))
+            XCTAssertEqual(error.localizedDescription, "\(emptyFileURL) is not a file")
         }
 
         let directoryURL = URL(fileURLWithPath: "./").absoluteURL   // file:///private/tmp/
         XCTAssertThrowsError(try FileLogger("com.example.yourapp.filelogger.notfile1", fileURL: directoryURL)) { error in
             XCTAssertEqual(error as? FileError, .isNotFile(url: directoryURL))
+            XCTAssertEqual(error.localizedDescription, "\(directoryURL) is not a file")
         }
     }
+
+    func testFilePermission() throws {
+        let fileURL = URL(fileURLWithPath: "./permission600.log").absoluteURL
+        let fileLogger = try FileLogger("com.example.yourapp.filelogger.permission600", fileURL: fileURL, filePermission: "600")
+        let log = Puppy()
+        log.add(fileLogger)
+        log.trace("permission, TRACE message using FileLogger")
+        log.verbose("permission, VERBOSE message using FileLogger")
+        fileLogger.flush()
+
+        let attribute = try FileManager.default.attributesOfItem(atPath: fileURL.path)
+        // swiftlint:disable force_cast
+        let permission = attribute[FileAttributeKey.posixPermissions] as! UInt16
+        // swiftlint:enable force_cast
+
+        #if os(Windows)
+        // NOTE: If the file is writable, its permission is always "700" on Windows.
+        // Reference: https://github.com/apple/swift-corelibs-foundation/blob/main/Sources/Foundation/FileManager%2BWin32.swift
+        let expectedPermission = UInt16("700", radix: 8)!
+        #else
+        let expectedPermission = UInt16("600", radix: 8)!
+        #endif // os(Windows)
+        XCTAssertEqual(permission, expectedPermission)
+
+        _ = fileLogger.delete(fileURL)
+        log.remove(fileLogger)
+    }
+
+    func testFilePermissionError() throws {
+        let permission800FileURL = URL(fileURLWithPath: "./permission800.log").absoluteURL
+        let filePermission800 = "800"
+        XCTAssertThrowsError(try FileLogger("com.example.yourapp.filelogger.permission800", fileURL: permission800FileURL, filePermission: filePermission800)) { error in
+            XCTAssertEqual(error as? FileError, .invalidPermission(at: permission800FileURL, filePermission: filePermission800))
+            XCTAssertEqual(error.localizedDescription, "invalid file permission. file: \(permission800FileURL), permission: \(filePermission800)")
+        }
+
+        let permissionABCFileURL = URL(fileURLWithPath: "./permissionABC.log").absoluteURL
+        let filePermissionABC = "ABC"
+        XCTAssertThrowsError(try FileLogger("com.example.yourapp.filelogger.permissionABC", fileURL: permissionABCFileURL, filePermission: filePermissionABC)) { error in
+            XCTAssertEqual(error as? FileError, .invalidPermission(at: permissionABCFileURL, filePermission: filePermissionABC))
+            XCTAssertEqual(error.localizedDescription, "invalid file permission. file: \(permissionABCFileURL), permission: \(filePermissionABC)")
+        }
+    }
+
+    #if canImport(Darwin)
+    func testWritingError() throws {
+        let fileURL = URL(fileURLWithPath: "./readonly.log").absoluteURL
+        XCTAssertThrowsError(try FileLogger("com.example.yourapp.filelogger.readonly", fileURL: fileURL, filePermission: "400")) { error in
+            XCTAssertEqual(error as? FileError, .openingForWritingFailed(at: fileURL))
+            XCTAssertEqual(error.localizedDescription, "failed to open a file for writing: \(fileURL)")
+            // swiftlint:disable force_try
+            try! FileManager.default.removeItem(at: fileURL)
+            // swiftlint:enable force_try
+        }
+    }
+    #endif // canImport(Darwin)
 
     #if canImport(Darwin)
     func testCreatingError() throws {
@@ -64,12 +122,14 @@ final class FileLoggerTests: XCTestCase {
         XCTAssertThrowsError(try FileLogger("com.example.yourapp.filelogger.notcreatedirectory",
                                             fileURL: fileURLNotAbleToCreateDirectory)) { error in
             XCTAssertEqual(error as? FileError, .creatingDirectoryFailed(at: directoryURLNotAbleToCreateDirectory))
+            XCTAssertEqual(error.localizedDescription, "failed to create a directory: \(directoryURLNotAbleToCreateDirectory)")
         }
 
         let fileURLNotAbleToCreateFile = URL(fileURLWithPath: "/foo.log").absoluteURL   // file:///foo.log
         XCTAssertThrowsError(try FileLogger("com.example.yourapp.filelogger.notcreatefile",
                                             fileURL: fileURLNotAbleToCreateFile)) { error in
             XCTAssertEqual(error as? FileError, .creatingFileFailed(at: fileURLNotAbleToCreateFile))
+            XCTAssertEqual(error.localizedDescription, "failed to create a file: \(fileURLNotAbleToCreateFile)")
         }
     }
     #endif // canImport(Darwin)
@@ -84,7 +144,7 @@ final class FileLoggerTests: XCTestCase {
         case .success(let url):
             XCTAssertEqual(existentFileURL, url)
         case .failure:
-            XCTFail("shuould not be failed, but was failed")
+            XCTFail("should not be failed, but was failed")
         }
 
         let resultFailure = fileLogger.delete(noExistentFileURL)
@@ -92,8 +152,8 @@ final class FileLoggerTests: XCTestCase {
         case .success:
             XCTFail("should not be successful, but was successful")
         case .failure(let error):
-            XCTAssertEqual(error as FileDeletingError, .failed(at: noExistentFileURL))
-            XCTAssertEqual(error.localizedDescription, "failed to delete the file: \(noExistentFileURL)")
+            XCTAssertEqual(error as FileError, .deletingFailed(at: noExistentFileURL))
+            XCTAssertEqual(error.localizedDescription, "failed to delete a file: \(noExistentFileURL)")
         }
     }
 
@@ -109,7 +169,7 @@ final class FileLoggerTests: XCTestCase {
                 XCTAssertEqual(existentFileURL, url)
                 expSuccess.fulfill()
             case .failure:
-                XCTFail("shuould not be failed, but was failed")
+                XCTFail("should not be failed, but was failed")
             }
         }
 
@@ -119,8 +179,8 @@ final class FileLoggerTests: XCTestCase {
             case .success:
                 XCTFail("should not be successful, but was successful")
             case .failure(let error):
-                XCTAssertEqual(error as FileDeletingError, .failed(at: noExistentFileURL))
-                XCTAssertEqual(error.localizedDescription, "failed to delete the file: \(noExistentFileURL)")
+                XCTAssertEqual(error as FileError, .deletingFailed(at: noExistentFileURL))
+                XCTAssertEqual(error.localizedDescription, "failed to delete a file: \(noExistentFileURL)")
                 expFailure.fulfill()
             }
         }
@@ -144,8 +204,8 @@ final class FileLoggerTests: XCTestCase {
             _ = try fileLogger.delete(noExistentFileURL).get()
             XCTFail("error should be thrown while awaiting, but it was not thrown")
         } catch {
-            XCTAssertEqual(error as? FileDeletingError, .failed(at: noExistentFileURL))
-            XCTAssertEqual(error.localizedDescription, "failed to delete the file: \(noExistentFileURL)")
+            XCTAssertEqual(error as? FileError, .deletingFailed(at: noExistentFileURL))
+            XCTAssertEqual(error.localizedDescription, "failed to delete a file: \(noExistentFileURL)")
         }
     }
 
@@ -171,8 +231,8 @@ final class FileLoggerTests: XCTestCase {
                 _ = try result.get()
                 XCTFail("error should be thrown while awaiting, but it was not thrown")
             } catch {
-                XCTAssertEqual(error as? FileDeletingError, .failed(at: noExistentFileURL))
-                XCTAssertEqual(error.localizedDescription, "failed to delete the file: \(noExistentFileURL)")
+                XCTAssertEqual(error as? FileError, .deletingFailed(at: noExistentFileURL))
+                XCTAssertEqual(error.localizedDescription, "failed to delete a file: \(noExistentFileURL)")
                 expFailure.fulfill()
             }
         }
@@ -212,44 +272,5 @@ final class FileLoggerTests: XCTestCase {
         }
 
         wait(for: [exp], timeout: 5.0)
-    }
-
-    func testFilePermission() throws {
-        let fileURL = URL(fileURLWithPath: "./permission600.log").absoluteURL
-        let fileLogger = try FileLogger("com.example.yourapp.filelogger.permission600", fileURL: fileURL, filePermission: "600")
-        let log = Puppy()
-        log.add(fileLogger)
-        log.trace("permission, TRACE message using FileLogger")
-        log.verbose("permission, VERBOSE message using FileLogger")
-        fileLogger.flush()
-
-        let attribute = try FileManager.default.attributesOfItem(atPath: fileURL.path)
-        // swiftlint:disable force_cast
-        let permission = attribute[FileAttributeKey.posixPermissions] as! UInt16
-        // swiftlint:enable force_cast
-
-        #if os(Windows)
-        // NOTE: If the file is writable, its permission is always "700" on Windows.
-        // Reference: https://github.com/apple/swift-corelibs-foundation/blob/main/Sources/Foundation/FileManager%2BWin32.swift
-        let expectedPermission = UInt16("700", radix: 8)!
-        #else
-        let expectedPermission = UInt16("600", radix: 8)!
-        #endif // os(Windows)
-        XCTAssertEqual(permission, expectedPermission)
-
-        _ = fileLogger.delete(fileURL)
-        log.remove(fileLogger)
-    }
-
-    func testFilePermissionError() throws {
-        let permission800FileURL = URL(fileURLWithPath: "./permission800.log").absoluteURL
-        XCTAssertThrowsError(try FileLogger("com.example.yourapp.filelogger.permission800", fileURL: permission800FileURL, filePermission: "800")) { error in
-            XCTAssertEqual(error as? FileError, .invalidPermission(at: permission800FileURL, filePermission: "800"))
-        }
-
-        let permissionABCFileURL = URL(fileURLWithPath: "./permissionABC.log").absoluteURL
-        XCTAssertThrowsError(try FileLogger("com.example.yourapp.filelogger.permissionABC", fileURL: permissionABCFileURL, filePermission: "ABC")) { error in
-            XCTAssertEqual(error as? FileError, .invalidPermission(at: permissionABCFileURL, filePermission: "ABC"))
-        }
     }
 }