+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Remove concurrency features. [#39](https://github.com/sushichop/Puppy/pull/39)
 - Fix the default rotation type to `numbering`. [#40](https://github.com/sushichop/Puppy/pull/40)
 - Change the type of intPermission to `UInt16`. [#41](https://github.com/sushichop/Puppy/pull/41)
+- Output a message prompting to override. [#42](https://github.com/sushichop/Puppy/pull/42)
 
 ## [0.4.0](https://github.com/sushichop/Puppy/releases/tag/0.4.0) (2022-01-31)